walker weight initialization
